pyproject.toml
[tool.poetry]
name = "gocq-py"
version = "0.1.0"
description = ""
authors = ["copurxia <copur@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gocq_py"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^2.2.3"}
loguru = "^0.6.0"
gevent = "^22.10.2"
pyyaml = "^6.0"
requests = "^2.28.2"
edgegpt = "^0.0.54"
revchatgpt = "^3.0.9"
tls-client = "^0.1.8"
pymongo = "^4.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
